refactor(frame): split overlay responsibilities
